+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Tokenizer ability to parse "@include(" into Token::Include, Token::OpenParen.
+  Added token UnescapedAt. 6 tests for @include cases.
 - Tokenizer.parse_error_prev.
 - Implemented Tokenizer.return_tokenized.
 - Renamed Tokenizer.fail to fail_token and refactred it so that it can be